feat: style the footer
Dividers were added between sections to better differentiate between them.

The logo was exchanged for a different layout to match the height of the menus better.

Then footer was centered to match the rest of the page.

Brand colors were applied and links added where appropiate. Padding was added for better visual appeal.

The menus were styled with better legibility in mind.