feat: add and style content of empty cart page
Links to go back to shopping and to log in were added.

The content was centered to align with the header and footer.

Padding was added and the headings styled to differentiate between sections.

The button and the link were styled to be recognized as such.

Brand colors were added for a cohesive look.