feat: add and style content of the login page
The login form with the options to reset the password and to create an account were added.

The content was centered to match the header and footer.

Padding was added for better visual effect and the brand colors were applied for a cohesive look.

The button and the links were styled to be recognized as such.

The input fields received placeholders to signal the user which input is expected.