feat: import product catalog from JSON file to homepage
The JSON file was imported and the products were mapped on the homepage, while differentiating
between featured and non-featured products.

For each product card, the product image, name and price were imported and displayed.
The option was implemented for the user to change the amount of each item
as well as an 'add to cart' button. The user may increase or decrease the amount by one or input
a number of their choice.