feat: add functionality to quantity counter
When using the plus and minus button the amount is increased or decreased by one and displayed as such.
The user might input an amount directly into the input field.
The amount is stored in the respective "add to cart" button for further use.

Since those are handcrafted products, only an amount up to 30 products may be purchased.
So it is not possible to increase the amount higher than that or decrease it below 1.
A notification with instructions pops up, should the user attempt to input an invalid value.