feat: add a visible counter to the cart icon
A counter, located at the cart icon in the header, displays the amount of items in the cart, if there are any.
Otherwise the counter is not visible. This privides a better overview for the user of how many items (not different
products) are located in the cart at any given time.

Brand colors have been applied for a cohesive look.