feat: create and style container for products in cart
A container for the display of the products in cart was created.
It contains the possibility to return to the homepage and continue shopping.
An overview was created to differentiate between product details, amount and subprice.
A field to display the total price for all products was added as well as the purchase button.
The fields were style with padding and brand colors for better visual appeal and cohesiveness.