feat: import and style products from local storage on the cart page
The cart object in the local storage was imported to the cart page and styled in accordance
to the rest of the website.
The product images, name, prices and quantities were displayed in the page.
A button to remove the item from cart was implemented as well as
the display of the total price per item.
The amount of products was styled in such a way that the user will be able to
change the quantity of the items in the cart.

When the cart is empty it says so on the cart page,
otherwise it shows the list of products.

The display of the total price for all products was implemented.