feat: enable manipulation of product amount in cart
The plus buttons increase the amount of products by one, the minus buttons decrease the amount by one.
The amount is displayed accordingly, as well as the subtotal per product and the total price.

The user may enter an amount between 1 and 30 that will be logged and processed.
Should the user enter an invalid token, an alert pops up with instructions.