feat: add option to create account
A register page was created to create a new user account that can be accessed through the
login page.
The form checks the email address, whether all fields have been filled in and whether the
email address has been registered with before.

Upon successful registration, the user data gets stored in the local storage and the user gets
forwarded to the homepage.