feat: create option to login
When the user uses the login form, the program checks whether email and password
match with the database. It also checks whether the user is already logged in.
Upon successful login the user is forwarded to the homepage. Moreover, a personalized
greeting appears in the header along with the option to log out (on all pages).

The login information is stored in local storage and is deleted when the user logs out.
Then the greeting and the logout button disappear again till the next login.