feat: implement process upon purchase
If the purchase button is clicked, the program checks whether the user is logged in.
Is this not the case, then the user is being forwarded to the login page.

If the user is logged in, the order along with the user's email is stored in local storage.
The user is forwarded to the "thank you" page.
The cart is emptied and the number on the cart icon is reset and hidden.