feat: implement search feature
If the user inputs a search query into the search bar, a dropdown with products appears that
contain the search term in their name.
The products are being rendered via JavaScript onto the page with the image and name of the
product. The dropdown disappears, once the search field is empty.

When the user hovers with the mouse over a product in the search results, the product is
highlighted for better visibility.

On the homepage, when a product in the search results is clicked on, the product scrolls into view.
On any other page, the user is redirected to the homepage.

Once a product in the search results has been clicked on, the input field empties and the search form closes.