feat: add a review form to products
If any of the "write a review" links is clicked, a review form appears, so the user can submit
a product review.
The product ID of the respective product gets stored in the local storage for further use.
If the user clicks the cancel button, the review form closes and the product ID is
deleted from the local storage.

The review form was styled according to the rest of the website. Brand colors
and padding were applied for better visual appeal and cohesiveness.
The respective product image is displayed within the review form.
A star rating system and text input have been incorporated into the form.