feat: provide functionality to review form
The user can rate the product by clicking on one of the stars in the
review form. Upon hover the stars change color to symbolize the rating
that will be logged when the star is clicked.
If the star is clicked, the hover effect is disabled. It is still possible to change
rating until the review is submitted. Once clicked, the rating is stored in the
local storage.
The star rating is obligatory.

Inside the input field, the user may leave a written review. Once entered, the
review is stored inside the local storage as well. It is still possible to change
the review until the form is submitted.

Upon cancel, all the stored data is deleted from the local storage.

If submit is clicked, the corresponding product id, the star rating and the written
review, if there is one, are bundled together and stored in local storage.
The previously logged review data is deleted, the review form is reset and the
review form window closed.